pull out shared board form component
